third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpotLightElement-svgdom-limitingConeAngle-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'limitingConeAngle' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.limitingConeAngle.baseVal is 45
PASS spotLightElement.limitingConeAngle.baseVal is 15
PASS successfullyParsed is true

TEST COMPLETE
